services/processor/requirements.txt
# ElevatedIQ News Feed Processor
# Python-based content analysis using Claude AI

# Core AI dependencies
anthropic>=0.40.0
openai>=1.56.0

# Configuration and validation
pydantic>=2.7.0
pydantic-settings>=2.2.0

# Database clients
psycopg[binary]>=3.2.0
psycopg_pool>=3.2.0
redis>=5.0.0
asyncpg>=0.30.0

# Message queue
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# Numerical processing
numpy>=1.26.4
scipy>=1.12.0
statsmodels>=0.14.0
scikit-learn>=1.4.0

# HTTP clients
httpx>=0.28.0
aiohttp>=3.9.4

# Content parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.3.0

# Date/time handling
python-dateutil>=2.9.0
pytz>=2024.1

# Logging and monitoring
structlog>=24.4.0
prometheus-client>=0.21.0

# Retry logic
tenacity>=9.0.0

# Social media integrations
youtube-transcript-api>=0.6.2
tweepy>=4.14.0
praw>=7.7.1
pytrends>=4.9.0

# GCP integration
google-cloud-secret-manager>=2.18.0

# Encryption and security
cryptography>=42.0.0

# Video generation
elevenlabs>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
mypy>=1.8.0
ruff>=0.3.0

# Type stubs
types-redis>=4.6.0
types-pytz>=2024.1.0
